fix: dataType int

+++ src/database/migrations/20230917085030-events.js
@@ -13,7 +13,7 @@ module.exports = {
      */
     queryInterface.createTable(tableName, {
       id: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         primaryKey: true,
         autoIncrement: true,
       },
@@ -22,11 +22,11 @@ module.exports = {
         allowNull: false,
       },
       formatId: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       topicId: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       coverUrl: {
